\name{posnegRichardsInit}
\alias{posnegRichardsInit}
\title{posnegRichardsInit}
\description{Not a user accessible function - integral to the workings of selfStart 
         \code{\link{SSposnegRichards}}}
\details{Not a user accessible function - integral to the workings of selfStart 
         \code{\link{SSposnegRichards}}
}
\usage{
posnegRichardsInit(mCall,
LHS,
data,
modno)
}
\arguments{
  \item{mCall}{ Call from \code{\link{SSposnegRichards}}
}
  \item{LHS}{ Left hand side of the formula from \code{\link{SSposnegRichards}}
}
  \item{data}{ Data object specified from \code{\link{SSposnegRichards}}
}
  \item{modno}{ Identification number of the model to be fitted 
     specified from \code{\link{SSposnegRichards}}
}
}

\seealso{\code{\link{SSposnegRichards}}}

\author{Stephen Oswald <ozsao23@hotmail.com>}





